Organize toolbox tests the same way as toolbox.
Previously these utilities were all in app/utils, but now that it's been
moved to toolbox and split up, there is no toolbox/utils, but rather
toolbox/{discord,github,messages,misc}, etc.
